ability to filter by deployment environment service map (#2506)
