+++ frontend/src/store/actions/serviceMap.ts
@@ -1,4 +1,6 @@
 import api from 'api';
+import { IResourceAttribute } from 'hooks/useResourceAttribute/types';
+import { convertRawQueriesToTraceSelectedTags } from 'hooks/useResourceAttribute/utils';
 import { Dispatch } from 'redux';
 import { GlobalTime } from 'types/actions/globalTime';
 
@@ -30,16 +32,17 @@ export interface ServiceMapLoading {
 	};
 }
 
-export const getDetailedServiceMapItems = (globalTime: GlobalTime) => async (
-	dispatch: Dispatch,
-): Promise<void> => {
+export const getDetailedServiceMapItems = (
+	globalTime: GlobalTime,
+	queries: IResourceAttribute[],
+) => async (dispatch: Dispatch): Promise<void> => {
 	const start = `${globalTime.minTime}`;
 	const end = `${globalTime.maxTime}`;
 
 	const serviceMapPayload = {
 		start,
 		end,
-		tags: [],
+		tags: convertRawQueriesToTraceSelectedTags(queries),
 	};
 	const [dependencyGraphResponse] = await Promise.all([
 		api.post<ServicesMapItem[]>(`/dependency_graph`, serviceMapPayload),
